⬆️ Update rust to 1.47.0
Rust Analyzer is now shipped with the distribution in the Mozilla
nixpkgs overlay so we just need to make sure rust-src is available in
the shell. This will also be more correct since the source will actually
match what we build against (not the case when using rust-analyzer from
nixpkgs).