✨ Add rust vendoring
This adds two derivations that rust packages can depend on, one for
internal dependencies and one for external. These derivations create
what cargo calls vendor directories. These can be searched by Cargo
instead of crates.io. We use cargo vendor to create a derivation for
crates.io dependencies. However, cargo does not really support creating
a vendor directory with unpublished crates (like we need to for internal
dependencies). Therefore, we have reverse-engineered the format of a
cargo vendor directory and create one for our internal dependencies by
simply unpacking the .crate file that cargo generates (a tar.gz) and
then generate the needed .cargo-checksum.json in the unpacked directory.

After those two derivations have been created, we use setup hooks in nix
to generate a .cargo/config.toml that redirects cargo to them. For our internal
vendor directory we use a fake registry that is then redirected to a
local folder. Cargo seems to accept this even though it is said to be
unsupported.

To Nedryland rust packages, this means that they can depend on internal
rust packages by specifying them in `buildInputs` or
`propagatedBuildInputs` and then adding a `Cargo.toml` dependency with
the registry set to "nix" like this:

    firm-types = { version = "0.1.0", registry = "nix" }

Currently we do not support if you have your own Cargo config at
.cargo/config.toml and will tell you to remove it if it is there.
Inclusion of custom config may be added later if/when needed.

Furthermore, at the moment vendoring of external dependencies happens
only for applications, not for libraries. This is since the normal build
artifact of a rust library is the source code for it, i.e. it is not
built. We do however build libraries in their check phases so we could
look into using vendoring there.