✨ Add languages helpers for protobuf
`languages.protobuf.generateCode` creates a derivation with protobuf
generated code for each language passed in the `languages` list.

`languages.protobuf.mkModule` creates nedryland components from the
generated code in a sensible way for each language passed in the
`languages` list. In contrast to the derivations above they can be used,
as is, in a nedryland build matrix.