🐛 Remove false runtime dependencies for rust
Store paths to the nix store end up in the rust binary. Nix tries to
be helpful by automatically adding dependencies to the package if it
finds a path that is matching something in the store.

This is not the case. Rust just has references to these paths to print
out good error messages.

Now we strip these store paths in the generated binary which fixes the problem.