🔥 Remove vendoring of external dependencies
Internal dependencies are still "vendored" like before.

It adds unneccesary complexity at this point in the project and cargo
vendor, with the requirements it has on the up-to-dateness of
Cargo.lock, makes vendoring inconvenient.

Instead, we should use Cargo.lock as an input to create a series of
fixed output derivations (the hash is in Cargo.lock) followed by a
derivation to create a vendor directory from these.