🐛 Change test command name from test to check
test is a command used in bash for if statements. When the user ran
other commands with tab completion etc it could suddenly start running
our tests which was not expected. It was because it was enough that
the auto completion script had an if statement inside it which would
trigger running our tests instead.