🔨 Refactor grid -> matrix and remove explicit mkMatrix
This consolidates matrix creation with project creation. Components
for a project is now sent in the `components` key to
`mkProject`. Furthermore, any key sent to `mkProject` can be a
function accepting a set as argument. This set is then intersected
with `base` (the unextended version) so any function from there is
available to use.

`declareComponent` is replaced with the supercharged function
`callFile` which will automatically send in any set argument named as
a package in nixpkgs or a component in the project. This is amazing!

`callFile` is also part of `base` so declaring nested components
should be easy-peasy now.

`projectDependencies` is also removed and users are instead expected
to put in the base extensions from projects that they want in the call
to `mkProject`.

`configFile` is now optional.

Any extra targets for the build matrix are now accepted as attributes
to `mkProject` and has the same possibility of being a function
as described above.