✨ Add windows rust runner
This will download a VM image and start a virtual machine running
windows. Cargo run and cargo test will then use ssh to run tests and
to run the executable on Windows. Note that the download of the VM
runs outside of nix and is therefore stored in /var/cache/nedryland
if the folder is writable and else in ~/.cache/nedryland.