🐛 Fix shell not getting check inputs
Now we instead use the same code that enable checks on the matrix to
enable it on the components individually. The reason that the shells
do not just use the matrix variant with checks is that it would cause
all inputs to all shells to run with checks, something that a user
would not expect.