✨ Add target setup to shells
Python and rust uses `mkTargetSetup` which takes a directory of files to
be substituted into a new component, along with variables to substitute
and commands to run. This is run if the any of the passed in marker files
are missing. It can prompt users and will read settings under the
`component` key in the project's config.