✨ Add `docs` target and markdown as language
The markdown language has functions to convert it to documentation using
different generators. Components can have a `docs` target with such
(or other) derivations. For python and rust this target gets merged with
api reference generated from the source code.

```nix
  mkComponent {
    ...
    docs = {
      userManual = base.languages.markdown.mkDocs { src = ./manual;};
    };
  };
```

This commit uses branches on sphinx and rtd_sphinx_theme to fix support
for logo as url which was introduced in sphinx 4.0 but did not work
properly. Logo as a path has to be absolute at the moment, because a
relative path will be evaluated to something unexpected.

This commits adds the attribute `nedrylandType` to components. This is
added so the config can put different requirements on different
component types, and also automatically work with component types
created by extensions that sets the type.