✨ Make config handle paths
Relative paths in nix are a bit tricky since they are relative to the
nix file executing them and once you are inside a buildPhase nix cuts
the ties to the outside world. With this we interpret a string starting
with "./" to be a path relative to the project's config file. This is
using `builtins.path` in order to make the file invariant of where it is
actually located.