🔥 Remove ability to download a Windows VM
This was used as an automated way of running Rust code on a Windows
machine. The usefulness of the feature did simply not warrant the
complexity. Instead, you can set `NEDRYLAND_WINDOWS_HOST` to a machine
where you have passwordless SSH access and cargo will use that machine
for `run/test`.