⚡ Add support building cross only for rust
By default the package builds on the host platform. Before we always
got this default host build. This does not always make sense since
some libraries or the like may only build on for example windows and
it does not make sense to have them build for any other os.

- Add support for rust to only build for specific cross targets and
not build host platform.
- Add the same feature for docs.
- Deduplicated code between mkLibrary and mkComponentWith
- Rust attribute for rust components now contains all targets in a list.