✨ Simplify dependency declaration for python
Before when creating a python component you had to specify a function
that returned a list and that list had to contain the packages of
components. It would look like the following example.

```
{ myPyPackage }:
  base.languages.python.mkLibrary {
  ...
  buildInputs = (pythonPkgs: [ myPyPackage.package pythonPkgs.su ])
  ...
  }
```

You had to add .package on all internal packages which was annoying.
You no longer need to do that. Can specify the component directly.
You can also just specify a list instead of a function.

This change is applied to buildInputs, nativeBuildInputs, checkInputs
and propagatedBuildInputs for python components.