⬆️ Update nixpkgs 21.11
- Use the python-language-server in pkgs

- Had to change from pythonPkgs.pyls-mypy to pythonPkgs.pylsp-mypy
because the old package brought the deprecated PLS which rendered you
unable to get a shell.

- Remove babel overlay since the one for py 38 and 39 has the 2.9.1
version already. If we want to keep it for 37 we must name it
something else so we don't override it for all other python
versions. If we keep it as is it will cause conflicts between the
version we have and the version another package will use (both use
same "version" but the nix hash will be different since they aren't
exactly the same).

- Remove themes argument to mkProject together with mkTheme. It was
never very useful.