🐛 Fix non existing default shells
There was a case where the component did not have a default target and
had more than one shell. In those cases it can't decide which shell to
pick. Made it create a shell that gives the user an error and
providing it with the available shells and then exiting the shell.