🐛 Fix doCheck for python derivations
Nixpkg's python defaults to running checks so we turned them off by
default. But that meant that with the new logic in enableChecksOverride
we can never turn them back on again. By removing the attribute from the
derivation we can use our own default after `buildPythonPackage` is done
with it.