🐛 🎨 🐎 Allow missing shellCommands and abort faster
Don't error if the derivation is missing `shellCommands`. Also use the
callPackage pattern better. Switched to using `builtins.abort` instead
of creating a shell and then immediately exit. Kept the `mkShell`,
otherwise it became too eager and exited on every shell for a component
without a default shell, even if a specific shell was requested.