⚡ 🐛 💎 Fix nix Python dependencies, 🍒 from in 6.3.0
Some python packages in nixpkgs are not discoverable by pip (missing
installinfo/egg_info) which causes problems for dependencies specified
in `install_requires`. This commit makes Python work more akin to C++,
where dependencies are specified once in Nix, to make sure they are
present at build time and once in CMake/Make to actually depend on
them. The python equivalent would be Nix to ensure prescence and
`setup.py` to declare usage. This way of solving the problem also
preserves the validity of the packages as normal Python packages.