⚡ Make mdbook more extendable and run better
Run the expected hooks on build and install and let users extend
buildInputs (when switching to nativeBuildInputs). Shadowed the run
shellCommand with a shell function, this is so a trap can be defined in
the shell to shut down mdbook when exiting the shell.

Also make it use ephemeral for the run command.