🔨 Change targets and all to be link farms
Previously, they were lists which is nicer when working with flakes and
also produces better output (result/<targetName> instead of
result-*/). Also, move the "targets" axis on the matrix to a sub-attrset
called `targets`. This causes evaluation to be a lot faster when asking
to build/shell components.

Also improved the ci scripts and renamed them to checks. Uses
$NEDRYLAND_CHECK_FILES to only look at the given files and a `check`
script to run all checks in the bin folder, which makes it easy to add
more checks by symlinkJoin.