✨ Make everything into derivations
Building a component with multiple targets now correctly yields a
derivation that when built contains links to the target outputs. This
makes component targets easier to inspect through the file system.

One requirement that arises for this to function is that every level in
the hierarchy of components needs to fulfill the requirement that it is
a derivation. One simple way to create a derivation is to use
`base.mkComponent` instead of returning a "naked" attrset.