🐛 Fix shell nativeBuildInputs priority
The priority was supposed to be:

1. nativeBuildInputs
2. checkInputs
3. shellInputs
4. shellCommands

but since `overrideAttrs` runs on the input arguments to `mkDerivation`,
not on the output of it, the logic to add `checkInputs` to
`nativeBuildInputs` was not run until the final shell derivation was
created. To fix this, the final shell derivation is overridden instead,
which causes the addition logic to have been applied.