🐛 Fix nested shells
Now, all non-leaf nodes in a component hierarchy are actual Nedryland
components with their type set to "component-set". This makes the tree
easier to deal with since we know that a leaf node is a derivation that
is not a Nedryland component and a non-leaf node is a Nedryland
component - always.