⚡ 🐎 Add passthru to components
The attrset passthru works like on mkDerivation, values inside are not
passed to the linkFarm but added to the result.

Change the order so users have the ability to set any attributes on
their component without having it being overwritten by the linkFarm. The
user is expected to be responsible enough to not make a target called
outPath for example.